Move BluetoothSessionController to a separate file
